# adlBackend/swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/adl-api/v1/cars":{"post":{"tags":["car-data-rest-controller"],"operationId":"saveADL","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarData"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/adl-api/v1/ready":{"get":{"tags":["car-data-rest-controller"],"operationId":"tellIngressEverythingIsReady","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/adl-api/v1/cars/{vin}":{"get":{"tags":["car-data-rest-controller"],"operationId":"retrieveADLByVin","parameters":[{"name":"vin","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/adl-api/v1/cars/{vin}/{numberOfEntries}":{"get":{"tags":["car-data-rest-controller"],"operationId":"retrieveADLByVin_1","parameters":[{"name":"vin","in":"path","required":true,"schema":{"type":"string"}},{"name":"numberOfEntries","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/adl-api/v1/alive":{"get":{"tags":["car-data-rest-controller"],"operationId":"tellIngressEverythingIsAlive","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"CarData":{"type":"object","properties":{"timeStamp":{"type":"string","format":"date-time"},"vin":{"type":"string"},"gpsPosition":{"$ref":"#/components/schemas/GpsPosition"},"mileage":{"$ref":"#/components/schemas/Mileage"},"fillLevels":{"$ref":"#/components/schemas/FillLevels"},"embt":{"type":"integer","format":"int32"},"lightOperatingHours":{"type":"number","format":"double"},"tirePressure":{"type":"number","format":"double"},"temperature":{"type":"integer","format":"int32"},"nodsa":{"type":"integer","format":"int32"},"nomi":{"type":"integer","format":"int32"},"sblc":{"type":"integer","format":"int32"}}},"FillLevels":{"type":"object","properties":{"fuel":{"type":"integer","format":"int32"},"coolant":{"type":"integer","format":"int32"},"brakeFluid":{"type":"integer","format":"int32"},"washingWater":{"type":"integer","format":"int32"}}},"GpsPosition":{"type":"object","properties":{"lat":{"type":"number","format":"double"},"lon":{"type":"number","format":"double"}}},"Mileage":{"type":"object","properties":{"city":{"type":"integer","format":"int32"},"autobahn":{"type":"integer","format":"int32"},"country":{"type":"integer","format":"int32"}}}}}}